+++ Panther/src/Context.cpp
@@ -556,17 +556,17 @@ namespace pcit::panther{
 						break; case SymbolProc::BuiltinSymbolKind::PANIC:
 							infos.emplace_back("\t> panic");
 						break; case SymbolProc::BuiltinSymbolKind::ARRAY_ITERABLE:
-							infos.emplace_back("\t> array.Iterable");
+							infos.emplace_back("\t> array.IIterable");
 						break; case SymbolProc::BuiltinSymbolKind::ARRAY_ITERABLE_RT:
-							infos.emplace_back("\t> array.IterableRT");
+							infos.emplace_back("\t> array.IIterableRT");
 						break; case SymbolProc::BuiltinSymbolKind::ARRAY_REF_ITERABLE_REF:
-							infos.emplace_back("\t> arrayRef.IterableRef");
+							infos.emplace_back("\t> arrayRef.IIterableRef");
 						break; case SymbolProc::BuiltinSymbolKind::ARRAY_REF_ITERABLE_REF_RT:
-							infos.emplace_back("\t> arrayRef.IterableRefRT");
+							infos.emplace_back("\t> arrayRef.IIterableRefRT");
 						break; case SymbolProc::BuiltinSymbolKind::ARRAY_MUT_REF_ITERABLE_MUT_REF:
-							infos.emplace_back("\t> arrayMutRef.IterableMutRef");
+							infos.emplace_back("\t> arrayMutRef.IIterableMutRef");
 						break; case SymbolProc::BuiltinSymbolKind::ARRAY_MUT_REF_ITERABLE_MUT_REF_RT:
-							infos.emplace_back("\t> arrayMutRef.IterableMutRefRT");
+							infos.emplace_back("\t> arrayMutRef.IIterableMutRefRT");
 					}
 				}
 					
@@ -3185,12 +3185,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// Iterator
+		// IIterator
 
 		const BaseType::ID iterator_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("Iterator"),
+				pthr_module.createString("IIterator"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -3198,7 +3198,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("Iterator", iterator_id);
+		pthr_module.createSymbol("IIterator", iterator_id);
 
 		{
 			const TypeInfo::ID iterator_type_id = this->type_manager.getOrCreateTypeInfo(TypeInfo(iterator_id));
@@ -3358,12 +3358,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// MutIterator
+		// IMutIterator
 
 		const BaseType::ID mut_iterator_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("MutIterator"),
+				pthr_module.createString("IMutIterator"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -3371,7 +3371,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("MutIterator", mut_iterator_id);
+		pthr_module.createSymbol("IMutIterator", mut_iterator_id);
 
 		{
 			const TypeInfo::ID mut_iterator_type_id = this->type_manager.getOrCreateTypeInfo(TypeInfo(mut_iterator_id));
@@ -3530,12 +3530,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// Iterable
+		// IIterable
 
 		const BaseType::ID iterable_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("Iterable"),
+				pthr_module.createString("IIterable"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -3543,7 +3543,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("Iterable", iterable_id);
+		pthr_module.createSymbol("IIterable", iterable_id);
 
 		{
 			const TypeInfo::ID iterable_type_id = this->type_manager.getOrCreateTypeInfo(TypeInfo(iterable_id));
@@ -3551,7 +3551,7 @@ namespace pcit::panther{
 			BaseType::Interface& iterable_type = this->type_manager.getInterface(iterable_id.interfaceID());
 
 
-			// func createIterator = (this) -> impl($$:@pthr.Iterator);
+			// func createIterator = (this) -> impl($$:@pthr.IIterator);
 			const BaseType::ID create_iterator_type_id = this->type_manager.getOrCreateFunction(
 				BaseType::Function(
 					evo::SmallVector<BaseType::Function::Param>{
@@ -3609,7 +3609,7 @@ namespace pcit::panther{
 
 
 
-			// func createIterator = (this mut) -> impl($$:@pthr.MutIterator);
+			// func createIterator = (this mut) -> impl($$:@pthr.IMutIterator);
 			const BaseType::ID create_mut_iterator_type_id = this->type_manager.getOrCreateFunction(
 				BaseType::Function(
 					evo::SmallVector<BaseType::Function::Param>{
@@ -3670,12 +3670,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// IterableRef
+		// IIterableRef
 
 		const BaseType::ID iterable_ref_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("IterableRef"),
+				pthr_module.createString("IIterableRef"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -3683,7 +3683,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("IterableRef", iterable_ref_id);
+		pthr_module.createSymbol("IIterableRef", iterable_ref_id);
 
 		{
 			const TypeInfo::ID iterable_ref_type_id = this->type_manager.getOrCreateTypeInfo(TypeInfo(iterable_ref_id));
@@ -3691,7 +3691,7 @@ namespace pcit::panther{
 			BaseType::Interface& iterable_ref_type = this->type_manager.getInterface(iterable_ref_id.interfaceID());
 
 
-			// func createIterator = (this) -> impl($$:@pthr.Iterator);
+			// func createIterator = (this) -> impl($$:@pthr.IIterator);
 			const BaseType::ID create_iterator_type_id = this->type_manager.getOrCreateFunction(
 				BaseType::Function(
 					evo::SmallVector<BaseType::Function::Param>{
@@ -3751,12 +3751,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// IterableMutRef
+		// IIterableMutRef
 
 		const BaseType::ID iterable_mut_ref_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("IterableMutRef"),
+				pthr_module.createString("IIterableMutRef"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -3764,7 +3764,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("IterableMutRef", iterable_mut_ref_id);
+		pthr_module.createSymbol("IIterableMutRef", iterable_mut_ref_id);
 
 		{
 			const TypeInfo::ID iterable_mut_ref_type_id =
@@ -3774,7 +3774,7 @@ namespace pcit::panther{
 				this->type_manager.getInterface(iterable_mut_ref_id.interfaceID());
 
 
-			// func createIterator = (this) -> impl($$:@pthr.MutIterator);
+			// func createIterator = (this) -> impl($$:@pthr.IMutIterator);
 			const BaseType::ID create_iterator_type_id = this->type_manager.getOrCreateFunction(
 				BaseType::Function(
 					evo::SmallVector<BaseType::Function::Param>{
@@ -3835,12 +3835,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// IteratorRT
+		// IIteratorRT
 
 		const BaseType::ID iterator_rt_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("IteratorRT"),
+				pthr_module.createString("IIteratorRT"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -3848,7 +3848,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("IteratorRT", iterator_rt_id);
+		pthr_module.createSymbol("IIteratorRT", iterator_rt_id);
 
 		{
 			const TypeInfo::ID iterator_type_id = this->type_manager.getOrCreateTypeInfo(TypeInfo(iterator_rt_id));
@@ -4008,12 +4008,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// MutIteratorRT
+		// IMutIteratorRT
 
 		const BaseType::ID mut_iterator_rt_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("MutIteratorRT"),
+				pthr_module.createString("IMutIteratorRT"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -4021,7 +4021,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("MutIteratorRT", mut_iterator_rt_id);
+		pthr_module.createSymbol("IMutIteratorRT", mut_iterator_rt_id);
 
 		{
 			const TypeInfo::ID mut_iterator_type_id =
@@ -4181,12 +4181,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// IterableRT
+		// IIterableRT
 
 		const BaseType::ID iterable_rt_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("IterableRT"),
+				pthr_module.createString("IIterableRT"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -4194,7 +4194,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("IterableRT", iterable_rt_id);
+		pthr_module.createSymbol("IIterableRT", iterable_rt_id);
 
 		{
 			const TypeInfo::ID iterable_rt_type_id = this->type_manager.getOrCreateTypeInfo(TypeInfo(iterable_rt_id));
@@ -4202,7 +4202,7 @@ namespace pcit::panther{
 			BaseType::Interface& iterable_rt_type = this->type_manager.getInterface(iterable_rt_id.interfaceID());
 
 
-			// func createIterator = (this) #rt -> impl($$:@pthr.IteratorRT);
+			// func createIterator = (this) #rt -> impl($$:@pthr.IIteratorRT);
 			const BaseType::ID create_iterator_type_id = this->type_manager.getOrCreateFunction(
 				BaseType::Function(
 					evo::SmallVector<BaseType::Function::Param>{
@@ -4260,7 +4260,7 @@ namespace pcit::panther{
 
 
 
-			// func createIterator = (this mut) #rt -> impl($$:@pthr.MutIteratorRT);
+			// func createIterator = (this mut) #rt -> impl($$:@pthr.IMutIteratorRT);
 			const BaseType::ID create_mut_iterator_type_id = this->type_manager.getOrCreateFunction(
 				BaseType::Function(
 					evo::SmallVector<BaseType::Function::Param>{
@@ -4321,12 +4321,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// IterableRefRT
+		// IIterableRefRT
 
 		const BaseType::ID iterable_rt_ref_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("IterableRefRT"),
+				pthr_module.createString("IIterableRefRT"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -4334,7 +4334,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("IterableRefRT", iterable_rt_ref_id);
+		pthr_module.createSymbol("IIterableRefRT", iterable_rt_ref_id);
 
 		{
 			const TypeInfo::ID iterable_rt_ref_type_id =
@@ -4344,7 +4344,7 @@ namespace pcit::panther{
 				this->type_manager.getInterface(iterable_rt_ref_id.interfaceID());
 
 
-			// func createIterator = (this) #rt -> impl($$:@pthr.IteratorRT);
+			// func createIterator = (this) #rt -> impl($$:@pthr.IIteratorRT);
 			const BaseType::ID create_iterator_type_id = this->type_manager.getOrCreateFunction(
 				BaseType::Function(
 					evo::SmallVector<BaseType::Function::Param>{
@@ -4404,12 +4404,12 @@ namespace pcit::panther{
 
 
 		//////////////////
-		// IterableMutRefRT
+		// IIterableMutRefRT
 
 		const BaseType::ID iterable_rt_mut_ref_id = this->type_manager.createInterface(
 			BaseType::Interface(
 				BuiltinModule::ID::PTHR,
-				pthr_module.createString("IterableMutRefRT"),
+				pthr_module.createString("IIterableMutRefRT"),
 				std::nullopt,
 				std::nullopt,
 				false,
@@ -4417,7 +4417,7 @@ namespace pcit::panther{
 			)
 		);
 
-		pthr_module.createSymbol("IterableMutRefRT", iterable_rt_mut_ref_id);
+		pthr_module.createSymbol("IIterableMutRefRT", iterable_rt_mut_ref_id);
 
 		{
 			const TypeInfo::ID iterable_rt_mut_ref_type_id =
@@ -4427,7 +4427,7 @@ namespace pcit::panther{
 				this->type_manager.getInterface(iterable_rt_mut_ref_id.interfaceID());
 
 
-			// func createIterator = (this) #rt -> impl($$:@pthr.MutIteratorRT);
+			// func createIterator = (this) #rt -> impl($$:@pthr.IMutIteratorRT);
 			const BaseType::ID create_iterator_type_id = this->type_manager.getOrCreateFunction(
 				BaseType::Function(
 					evo::SmallVector<BaseType::Function::Param>{